requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alabaster==0.7.16
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
astroid==2.15.8
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.15.0
billiard==3.6.4.0
black==23.1.0
cachetools==5.3.3
celery==5.2.7
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.2.1
crispy-bootstrap5==0.7
cron-descriptor==1.4.3
dataclasses-json==0.6.6
decorator==5.1.1
dill==0.3.8
distlib==0.3.8
Django==4.2.13
django-celery-beat==2.5.0
django-coverage-plugin==3.0.0
django-crispy-forms==2.0
django-debug-toolbar==3.8.1
django-environ==0.10.0
django-extensions==3.2.1
django-model-utils==4.3.1
django-redis==5.2.0
django-stubs==1.15.0
django-stubs-ext==5.0.0
django-timezone-field==6.1.0
docutils==0.19
exceptiongroup==1.2.1
executing==2.0.1
factory-boy==3.2.1
Faker==25.1.0
filelock==3.14.0
flake8==6.0.0
flake8-isort==6.0.0
frozenlist==1.4.1
fsspec==2024.5.0
huggingface-hub==0.23.0
hyperframe==6.0.1
Pillow==9.4.0
portalocker==2.8.2
pre-commit==3.1.1
prompt-toolkit==3.0.43
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycodestyle==2.10.0
pycparser==2.22
pydantic==2.7.1
pydantic-core==2.18.2
pyflakes==3.0.1
pygments==2.18.0
pylint==2.17.7
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
pyparsing==3.1.2
pytest==7.2.2
pytest-django==4.5.2
pytest-sugar==0.9.6
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.1
python-telegram-bot==21.1.1
pytz==2022.7.1
PyYAML==6.0.1
qdrant-client==1.9.1
redis==4.5.1
regex==2024.5.15
requests==2.31.0
rsa==4.9
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
sentence-transformers==2.7.0
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
sphinx==6.1.3
sphinx-autobuild==2021.3.14
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
SQLAlchemy==2.0.30
sqlparse==0.5.0
stack-data==0.6.3
sympy==1.12
tenacity==8.3.0
termcolor==2.4.0
text-unidecode==1.3
threadpoolctl==3.5.0
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.12.5
torch==2.3.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
transformers==4.40.2
types-pytz==2024.1.0.20240417
types-PyYAML==6.0.12.20240311
typing-extensions==4.11.0
typing-inspect==0.9.0
uritemplate==4.1.1
urllib3==2.2.1
google-generativeai
python-dotenv